Move DB init to own file, add app to docker-compose
Fjern metoder fra NaisRest
