+++ README.md
@@ -6,27 +6,22 @@ TODO
 * Kotlin
 * Ktor
 * Gradle
-* JDK 12
+* JDK 13
 * Spek
 * Jackson
 
 #### Requirements
-* Adopt-OpenJDK 12
+* Adopt-OpenJDK 13
 
 #### Build and run tests
-To build locally and run the integration tests you can simply run `./gradlew shadowJar` or on windows 
-`gradlew.bat shadowJar`
+To build locally and run the integration tests you can simply run `./gradlew test`
 
-#### Creating a docker image
-Creating a docker image should be as simple as `docker build -t ispengestopp .`
-
-#### Running a docker image
-`docker run --rm-it -p 8080:8080 ispengestopp`
-
+#### Running locally
+`docker compose up`
 
 ## Contact us
 ### Code/project related questions can be sent to
 * June Henriksen, `june.henriksen2@nav.no`
 * Mathias Rørvik, `mathias.fris.rorvik@nav.no`
 * John Martin Lindseth `john.martin.lindseth@nav.no`
-* The following channel on slack --> #isyfo 
+* The following channel on slack --> #isyfo
