+++ docker-compose.yml
@@ -1,5 +1,12 @@
 version: '3'
 services:
+  web:
+    build: .
+    ports:
+      - "8080:8080"
+    environment:
+      - DEVELOPMENT_MODE=true
+
   db:
     image: postgres
     ports:
@@ -8,4 +15,4 @@ services:
       - POSTGRES_DB=ispengestopp_dev
       - POSTGRES_USER=username
       - POSTGRES_PASSWORD=password
-      - DEVELOPMENT_MODE=true
+
