+++ src/main/kotlin/no/nav/syfo/api/NaisRest.kt
@@ -13,19 +13,17 @@ import no.nav.syfo.application.ApplicationState
 
 fun Routing.registerNaisApi(
     applicationState: ApplicationState,
-    readynessCheck: () -> Boolean = { applicationState.ready },
-    alivenessCheck: () -> Boolean = { applicationState.alive },
     collectorRegistry: CollectorRegistry = CollectorRegistry.defaultRegistry
 ) {
     get("/is_alive") {
-        if (alivenessCheck()) {
+        if (applicationState.alive) {
             call.respondText("I'm alive! :)")
         } else {
             call.respondText("I'm dead x_x", status = HttpStatusCode.InternalServerError)
         }
     }
     get("/is_ready") {
-        if (readynessCheck()) {
+        if (applicationState.ready) {
             call.respondText("I'm ready! :)")
         } else {
             call.respondText("Please wait! I'm not ready :(", status = HttpStatusCode.InternalServerError)
