+++ src/main/resources/logback.xml
@@ -7,7 +7,9 @@
     </appender>
 
     <logger name="no.nav" level="INFO"/>
-
+<!--    <root level="INFO">-->
+<!--        <appender-ref ref="STDOUT"/>-->
+<!--    </root>-->
     <root level="INFO">
         <appender-ref ref="stdout_json"/>
     </root>
